+++ api/v2/globalConfigDiff.go
@@ -1,6 +1,6 @@
-package v1
+package v2
 
-// GlobalConfigDiff is a list of differences between the GlobalConfig in the EffectiveBlueprint and the cluster state
+// GlobalConfigDiff is a list of differences between the GlobalConfig in the Blueprint and the cluster state
 type GlobalConfigDiff []GlobalConfigEntryDiff
 
 // GlobalConfigValueState represents either the actual or expected state of a global config key