+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+**Breaking Changes ahead**
+
+- [#5] **Breaking** use yaml format for blueprint and blueprint mask instead of json strings
+- [#5] **Breaking** use absent bool instead of targetState for dogus and components
+- [#5] **Breaking** only reference sensitive config by secret
+- [#5] update dependencies
+
 ## [v1.3.0] - 2025-06-04
 ### Added
 - [#3] blueprints now support additionalMounts for configMaps and secrets in dogus